requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.2
anyio==4.4.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
openai==1.51.2

